pyproject.toml
[project]
name = "mcp-server"
version = "0.4.0"
description = "Graphiti MCP Server"
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "mcp>=1.5.0",
    "neo4j>=5.26.0",
    "openai>=1.91.0",
    "instructor>=1.9.2",
    "diskcache>=5.6.3",
    "numpy>=1.0.0",
    "posthog>=3.0.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "graphiti-core>=0.17.3",
    "fastapi>=0.116.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "cryptography>=45.0.5",
    "apscheduler>=3.11.0",
]
